добавляет файлы модулей js
